+++ Projeto/Includes/light.cpp
@@ -2,7 +2,7 @@
 // Define o construtor da classe Light
 // O construtor recebe a posicao da luz e a intensidade
 // Os valores sao usados para inicializar os atributos da classe
-Light::Light(const glm::vec3& position, const glm::ivec3& intensity)
+Light::Light(const glm::vec3& position, const glm::vec3& intensity)
     : position(position), intensity(intensity) {}
 
 // Define o metodo getPosition da classe Light
@@ -14,6 +14,6 @@ glm::vec3 Light::getPosition() const {
 // Define o metodo getIntensity da classe Light 
 // O metodo retorna a intensidade da luz como um vetor de 3 dimensoes com valores inteiros
 // Os valores representam a intensidade da luz em cada canal de cor (RGB)
-glm::ivec3 Light::getIntensity() const {
+glm::vec3 Light::getIntensity() const {
     return intensity;
 }
